backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0

# Authentication & Security
Flask-JWT-Extended==4.5.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Serialization & Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Flask-Marshmallow==0.15.0

# Background Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# OSINT & PDF
reportlab==4.0.7
python-whois==0.8.0
dnspython==2.4.2
shodan==1.30.0
aiohttp==3.9.1

# MCP Integration
fastmcp==0.9.0
mcp==1.0.0

# Utilities
python-dotenv==1.0.0
pytz==2024.1
pydantic==2.5.2
requests==2.31.0
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
gevent==23.9.1